% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{age}
\usage{
age(birthdate, as_of_date = NULL)
}
\arguments{
\item{birthdate}{birthday in the form YMD}

\item{as_of_date}{current day in the form YMD, default is today}
}
\value{
age
}
\description{
Calculate someone's age based on their birthday and a provided date
}
