% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{change_firstlast_to_lastfirst}
\alias{change_firstlast_to_lastfirst}
\title{change_firstlast_to_lastfirst}
\usage{
change_firstlast_to_lastfirst(name)
}
\arguments{
\item{name}{The full name of someone in the form of "First_Name Last_Name"}
}
\value{
Last_Name, First Name
}
\description{
change first_name last_name, to last_name, first_name
This function does not work with two part last names as it assumes
there is only one space between the first and last name.
}
