% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_year.R
\name{change_school_year}
\alias{change_school_year}
\title{change_school_year}
\usage{
change_school_year(school_year)
}
\arguments{
\item{school_year}{a school year in the full form of 20XX-20XY}
}
\description{
Takes a value in the form 2016-2017
and converts to 2016-17
}
\examples{
 change_school_year("2014-2015")

}
\seealso{
\code{\link{change_school_year}} \code{\link{sy_form}} \code{\link{sy_number}}
}
