% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{cohort}
\alias{cohort}
\title{Cohort Tag}
\usage{
cohort(grade, school_year = "")
}
\arguments{
\item{grade}{Enrolled grade of the student  (0 - 12)}

\item{school_year}{School year either in 20XX-YY or 20XX (spring) format.}
}
\description{
Defines a student's cohort given the student's grade and current school year.

If you do not provide a school year the default is the current (spring) year.
e.g If today's date is Jan 10, 2017, the default year will be 2017,
however if the current date is Sep 10, 2017, the year will default to 2018.
}
\examples{
library(RUncommon)
cohort(10)
cohort(10, 2008)
}
