% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{colorgorical}
\alias{colorgorical}
\title{colorgorical}
\usage{
colorgorical(n = 10)
}
\description{
uses Colorgoical to create color paletes that work well.
Requires httr and jsonlite packages.
}
