% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_name_lookup.R
\name{create_colname_lookup}
\alias{create_colname_lookup}
\title{create_colname_lookup}
\usage{
create_colname_lookup(df)
}
\description{
This creates a lookup table from columns that are automatically 
named according to the scheme apple_1, apple_2, banana_3, banana_4 to
apple, apple_1, banana, banana_1
}
