% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eringrand.R
\docType{package}
\name{eringrand}
\alias{eringrand}
\alias{eringrand-package}
\title{eringrand is my personal supply of function that I use often enough not to want to rewrite everytime}
\description{
eringrand is my personal supply of function that I use often enough not to want to rewrite everytime
}
\keyword{internal}
