% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{len}
\alias{len}
\title{len}
\usage{
len(x)
}
\arguments{
\item{x}{vector}
}
\description{
For similarity between Python and R.  Returns the length of a vector.
}
\examples{
library(RUncommon)
x <- 1:5
len(x)
}
