% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.regression.or.ci.R
\name{logistic.regression.or.ci}
\alias{logistic.regression.or.ci}
\title{logistic.regression.or.ci}
\usage{
logistic.regression.or.ci(regress.out, level = 0.95)
}
\arguments{
\item{regress.out}{logistic regression output}

\item{level}{level to test at, default is 0.95}
}
\description{
This function takes the output from a glm
 (logistic model) command in R and provides not
 only the usual output from the summary command, but
 adds confidence intervals for all coefficients and OR's.
This version accommodates multiple regression parameters
}
