% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_stats.R
\name{mygini}
\alias{mygini}
\title{This calculates the gini coeff}
\usage{
mygini(pred, actual)
}
\arguments{
\item{pred}{vector of predictions}

\item{actual}{vector of actual values}
}
\description{
This calculates the gini coeff
}
