% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{read_clean_data}
\alias{read_clean_data}
\title{read_clean_data}
\usage{
read_clean_data(file, sheetname, ...)
}
\arguments{
\item{file}{The location of the file you are using.}

\item{...}{Additional parameters to pass to readxl}

\item{sheename}{The name of number of the sheet (as in readxl)}
}
\description{
Reads in clean data from excel using janitor package and readxl
- Read in the data, assumes na = "",
- Removes empty columns and rows
- Cleans column names
}
\examples{
\dontrun{
library(RUncommon)
/read_clean_data(filelocation, 1)
}
}
