% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{round_percent}
\alias{round_percent}
\title{round_percent}
\usage{
round_percent(x, dig = 1)
}
\arguments{
\item{x}{a numeric vector to format}

\item{dig}{the number of digits after the percent to round. Default to 1.}
}
\description{
A version of scales::percent() that allows for rounded digit
so that you can have 20% or 20.5% as needed.
}
\examples{
 round_percent(.24601)
 round_percent(.24601, dig = 0)
}
