% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUncommonR.R
\docType{package}
\name{RUncommon}
\alias{RUncommon}
\alias{RUncommon-package}
\title{uncommonR is my personal supply of functions that I use often enough not to want to rewrite everytime}
\description{
uncommonR is my personal supply of functions that I use often enough not to want to rewrite everytime
}
\keyword{internal}
