% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_year.R
\name{school_year_from_date}
\alias{school_year_from_date}
\title{school_year_from_date}
\usage{
school_year_from_date(date = lubridate::today())
}
\arguments{
\item{date}{date in the form of lubridate::today()}
}
\description{
Takes a date (in lubridate::today) format and converts to a given schools year.
}
\examples{
school_year_from_date()
}
\seealso{
\code{\link{change_school_year}} \code{\link{sy_form}} \code{\link{sy_number}}
}
