% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_smooth_func.R
\name{stat_smooth_func}
\alias{stat_smooth_func}
\title{stat_smooth_func}
\usage{
stat_smooth_func(
  mapping = NULL,
  data = NULL,
  geom = "smooth",
  position = "identity",
  ...,
  method = "auto",
  formula = y ~ x,
  se = TRUE,
  n = 80,
  span = 0.75,
  fullrange = FALSE,
  level = 0.95,
  method.args = list(),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  xpos = NULL,
  ypos = NULL
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{geom}{geom}

\item{position}{pos}

\item{...}{extra}

\item{method}{method}

\item{formula}{form}

\item{se}{se}

\item{n}{n}

\item{span}{span}

\item{fullrange}{full}

\item{level}{lev}

\item{method.args}{x}

\item{na.rm}{na.rm}

\item{show.legend}{y}

\item{inherit.aes}{z}

\item{xpos}{a}

\item{ypos}{b}
}
\description{
mostly taken from https://stackoverflow.com/questions/7549694/adding-regression-line-equation-and-r2-on-graph
}
\examples{
library(dplyr)
library(ggplot2)
iris \%>\%
  ggplot(aes(x = Sepal.Width, y = Petal.Width, color = Species)) +
  geom_point() +
  stat_smooth_func(geom = "text", method = "lm", parse = TRUE, hjust = 0) +
  facet_wrap(~Species)
}
