% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_stats.R
\name{rmse}
\alias{rmse}
\alias{mape}
\alias{mygini}
\alias{tpr}
\alias{tnr}
\alias{fpr}
\alias{fnr}
\alias{cls_acc}
\title{Statistics Functions}
\usage{
rmse(pred, actual)

mape(pred, actual)

mygini(pred, actual)

tpr(pred, actual)

tnr(pred, actual)

fpr(pred, actual)

fnr(pred, actual)

cls_acc(pred, actual)
}
\arguments{
\item{pred}{vector of predictions}

\item{actual}{vector of actual values}
}
\description{
Statistics Functions
}
\section{Functions}{
\itemize{
\item \code{rmse}: root mean standard error

\item \code{mape}: Mean Absolute Percent Error

\item \code{mygini}: gini coeff

\item \code{tpr}: True Positive Rate

\item \code{tnr}: True Negative Rate

\item \code{fpr}: False Positive Rate

\item \code{fnr}: False Negative Rate

\item \code{cls_acc}: Classification accuracy
}}

