% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_year.R
\name{sy_number}
\alias{sy_number}
\title{school_year}
\usage{
sy_number(school_year)
}
\description{
Takes a value in the form SY16-17 or 2016-17
and converts to a number representing the second/spring year.
}
\examples{
 sy_number("SY16-17")
}
