% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncommonR.R
\docType{package}
\name{uncommonR}
\alias{uncommonR}
\alias{uncommonR-package}
\title{uncommonR is my personal supply of functions that I use often enough not to want to rewrite everytime}
\description{
uncommonR is my personal supply of functions that I use often enough not to want to rewrite everytime
}
\keyword{internal}
