% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{yes_no}
\alias{yes_no}
\title{yes_no}
\usage{
yes_no(x)
}
\arguments{
\item{x}{vectors of 1s and 0s}
}
\description{
Changes all 1s to YES and 0s to NO.
Will not effect other values - so NAs remain the same.
}
