% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{change_firstlast_to_lastfirst}
\alias{change_firstlast_to_lastfirst}
\title{change first_name last_name, to last_name, first_name}
\usage{
change_firstlast_to_lastfirst(name)
}
\description{
change first_name last_name, to last_name, first_name
}
