% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{cohort}
\alias{cohort}
\title{Cohort Tag}
\usage{
cohort(grade, school_year = "")
}
\arguments{
\item{grade}{Enrolled grade of the student}

\item{school_year}{School year either in 20XX-YY or 20XX format.}
}
\description{
Cohort Tag
}
\examples{
library(RUncommon)
cohort(10)
cohort(10, 2008)
}
