% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colorgorical}
\alias{colorgorical}
\title{colorgorical}
\usage{
colorgorical(n = 10)
}
\arguments{
\item{n}{number of colors in the color palette. defaults to 10
Credit to \href{https://gist.github.com/slowkow/22daea426607416bfcd573ce9cbd89ab}{slowkrow}
 @export}
}
\description{
uses colorgorical to create color palettes that work well.
Requires httr and jsonlite packages.
}
