% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_many.R
\name{spread_many}
\alias{spread_many}
\title{spread_many}
\usage{
spread_many(
  data,
  key_col,
  ...,
  fill = NA,
  convert = TRUE,
  drop = TRUE,
  sep = "_"
)
}
\arguments{
\item{data}{data}

\item{key_col}{key_col}

\item{...}{List of column, unquoted to spread out}

\item{fill}{fill}

\item{convert}{convert}

\item{drop}{drop}

\item{sep}{sep}
}
\value{
data frame with new columns
}
\description{
Spread more than one column out at once
}
\examples{
library(dplyr)
library(tidyr)

fruits <- data.frame(
  case = c(1, 1, 2, 2),
  fruit = rep(c("Apple", "Orange"), 2),
  height = rnorm(4, 0, 1),
  width = rnorm(4, 0, 2)
)

spread_many(fruits, fruit, width, height)
}
